requirements.txt
absl-py>=0.11.0
anyio>=2.1.0
argh>=0.26.2
argon2-cffi>=20.1.0
async-generator>=1.10
attrs>=20.3.0
Babel>=2.9.0
backcall>=0.2.0
backports.functools-lru-cache>=1.6.1
bleach>=3.3.0
brotlipy>=0.7.0
certifi>=2020.6.20
cffi>=1.14.4
chardet>=4.0.0
click>=7.1.2
configparser>=5.0.1
contextlib2>=0.6.0.post1
cryptography>=2.9.2
cycler>=0.10.0
decorator>=4.4.2
defusedxml>=0.6.0
docker-pycreds>=0.4.0
entrypoints>=0.3
gitdb>=4.0.5
GitPython>=3.1.12
gql>=0.1.0
graphql-core>=3.1.3
idna>=2.10
importlib-metadata>=3.4.0
ipykernel>=5.4.3
ipython>=7.20.0
ipython-genutils>=0.2.0
jedi>=0.18.0
Jinja2>=2.11.3
joblib>=0.17.0
json5>=0.9.5
jsonschema>=3.2.0
jupyter-client>=6.1.11
jupyter-core>=4.7.1
jupyter-server>=1.3.0
jupyterlab>=3.0.7
jupyterlab-pygments>=0.1.2
jupyterlab-server>=2.2.0
kiwisolver>=1.3.1
llvmlite>=0.35.0
MarkupSafe>=1.1.1
matplotlib>=3.2.2
mistune>=0.8.4
mkl-fft>=1.2.0
mkl-random>=1.1.1
mkl-service>=2.3.0
ml-collections>=0.1.0
nbclassic>=0.2.6
nbclient>=0.5.1
nbconvert>=6.0.7
nbformat>=5.1.2
nest-asyncio>=1.4.3
notebook>=6.2.0
numba>=0.52.0
numpy>=1.19.2
nvidia-ml-py3>=7.352.0
olefile>=0.46
packaging>=20.9
pandas>=1.2.2
pandocfilters>=1.4.2
parso>=0.8.1
pathtools>=0.1.2
patsy>=0.5.1
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=8.1.0
pip>=20.3.3
prometheus-client>=0.9.0
promise>=2.3
prompt-toolkit>=3.0.14
protobuf>=3.13.0
psutil>=5.7.3
ptyprocess>=0.7.0
pycparser>=2.20
Pygments>=2.7.4
pyOpenSSL>=19.1.0
pyparsing>=2.4.7
pyrsistent>=0.17.3
PySocks>=1.7.1
python-dateutil>=2.8.1
pytz>=2021.1
PyYAML>=5.3.1
pyzmq>=20.0.0
requests>=2.25.1
scikit-learn>=0.23.2
scipy>=1.6.0
Send2Trash>=1.5.0
sentry-sdk>=0.19.5
setuptools>=52.0.0.post20210125
shortuuid>=1.0.1
six>=1.15.0
sktime>=0.5.3
smmap>=3.0.5
sniffio>=1.2.0
statsmodels>=0.12.2
subprocess32>=3.5.4
terminado>=0.9.2
testpath>=0.4.4
threadpoolctl>=2.1.0
torch>=1.7.1
torchaudio>=0.7.0a0+a853dff
torchinfo>=0.0.8
torchvision>=0.8.2
tornado>=6.1
traitlets>=5.0.5
typing-extensions>=3.7.4.3
urllib3>=1.26.3
wandb>=0.10.18
watchdog>=0.10.4
wcwidth>=0.2.5
webencodings>=0.5.1
wheel>=0.36.2
zipp>=3.4.0
torchmetrics>=0.3.1